% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_sites.R
\name{available_sites}
\alias{available_sites}
\title{AVAILABLE SITES
List all site and month combinations available in the cache}
\usage{
available_sites(cache_dir = "cache")
}
\arguments{
\item{cache_dir}{path to cache folder (default "cache")}
}
\value{
a data.frame with columns site_id and month
}
\description{
AVAILABLE SITES
List all site and month combinations available in the cache
}
