% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{CALCULATE INDICES
Calculate NDVI, NDMI, and BSI from Sentinel-2 bands}
\usage{
calc_indices(bands)
}
\arguments{
\item{bands}{named list returned by load_bands()}
}
\value{
a SpatRaster with three layers: NDVI, NDMI, BSI
}
\description{
CALCULATE INDICES
Calculate NDVI, NDMI, and BSI from Sentinel-2 bands
}
