% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_system_types.R
\name{compare_system_types}
\alias{compare_system_types}
\title{COMPARE SYSTEM TYPES
Compare index statistics between agrovoltaic system types}
\usage{
compare_system_types(results, sites_sf = agropvR::sites)
}
\arguments{
\item{results}{data.frame returned by load_site_results()}

\item{sites_sf}{the sites sf object included in the package (sites)}
}
\value{
a summary data.frame with mean and stdev per index per system type
}
\description{
COMPARE SYSTEM TYPES
Compare index statistics between agrovoltaic system types
}
