% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields}
\alias{fields}
\title{Reference agricultural field polygons}
\format{
an sf object with 25 rows and a geometry column (CRS: EPSG 25832)
}
\source{
https://agrivoltaicsmap.ise.fraunhofer.de/
}
\usage{
fields
}
\description{
Paired reference fields for each of the 25 agrovoltaic sites.
}
\keyword{datasets}
