% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_system_types.R
\name{list_system_types}
\alias{list_system_types}
\title{LIST SYSTEM TYPES
Show which sites are overhead vs interspace systems}
\usage{
list_system_types(sites_sf = agropvR::sites)
}
\arguments{
\item{sites_sf}{the sites sf object included in the package (sites)}
}
\value{
a data.frame with site_id and system_typ
}
\description{
LIST SYSTEM TYPES
Show which sites are overhead vs interspace systems
}
