% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bands.R
\name{load_bands}
\alias{load_bands}
\title{LOAD BANDS
Load and crop Sentinel-2 bands for a site}
\usage{
load_bands(scene, buffer_sf)
}
\arguments{
\item{scene}{STAC feature returned by search_sentinel()}

\item{buffer_sf}{polygon used to define the crop extent}
}
\value{
named list with SpatRaster elements: red, nir, swir, blue
}
\description{
LOAD BANDS
Load and crop Sentinel-2 bands for a site
}
