% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panels}
\alias{panels}
\title{Photovoltaic panel footprints}
\format{
An sf object with 655 rows and a geometry column (CRS: EPSG 25832)
}
\source{
Digitised in QGIS using satellite basemap imagery
}
\usage{
panels
}
\description{
Digitised polygons of individual photovoltaic panel arrays within
each agrovoltaic site. Multiple polygons per site, linked via site_id.
}
\keyword{datasets}
