% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_system_types.R
\name{plot_system_types}
\alias{plot_system_types}
\title{PLOT SYSTEM TYPES
Plot index comparison between agrovoltaic system types}
\usage{
plot_system_types(results, sites_sf = agropvR::sites)
}
\arguments{
\item{results}{data.frame returned by load_site_results()}

\item{sites_sf}{the sites sf object included in the package (sites)}
}
\value{
ggplot object
}
\description{
PLOT SYSTEM TYPES
Plot index comparison between agrovoltaic system types
}
