% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{DATA
Agrovoltaic facility polygons}
\format{
an sf object with 25 rows and a geometry column (CRS: EPSG 25832)
}
\source{
https://agrivoltaicsmap.ise.fraunhofer.de/
}
\usage{
sites
}
\description{
Digitised polygons of 25 agrovoltaic facilities in Bavaria and
Baden-Wuerttemberg, Germany.
}
\keyword{datasets}
