% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agropvR-package.R
\docType{package}
\name{agropvR-package}
\alias{agropvR}
\alias{agropvR-package}
\title{agropvR}
\description{
An R package to map the environmental impact of agrophotovoltaics across Baden-Wuerttemberg and Bavaria.
}
\author{
\strong{Maintainer}: Marlene Sehrbrock \email{first.last@example.com}

}
