% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cache.R
\name{download_cache}
\alias{download_cache}
\title{DOWNLOAD CACHE
Download pre-computed cache from Zenodo}
\usage{
download_cache(cache_dir = "cache", zenodo_id = "20705323")
}
\arguments{
\item{cache_dir}{Local folder to save files (default "data/cache")}

\item{zenodo_id}{Zenodo record ID (default is the agropvR dataset)}
}
\value{
Invisible NULL
}
\description{
Downloads all cached .rds and .tif files from Zenodo.
Only needs to be run once. Skips files already present.
}
