% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cache.R
\name{download_cache}
\alias{download_cache}
\title{DOWNLOAD CACHE
Download pre-computed cache from Zenodo}
\usage{
download_cache(
  cache_dir = "cache",
  zenodo_id = c("20713726", "20713763", "20713833")
)
}
\arguments{
\item{cache_dir}{local folder to save files (default "cache")}

\item{zenodo_id}{character vector of Zenodo record IDs}
}
\value{
invisible NULL
}
\description{
Downloads all cached .rds and .tif files from Zenodo.
Only needs to be run once. Skips files already present.
}
