% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_sentinel.R
\name{search_sentinel}
\alias{search_sentinel}
\title{SEARCH SENTINEL
Find the least cloudy Sentinel-2 scene for a site and month}
\usage{
search_sentinel(bbox_sf, month, limit = 20, max_cloud = max_cloud)
}
\arguments{
\item{bbox_sf}{object, its bounding box is used for the STAC search}

\item{month}{character string in "YYYY-MM" format, e.g. "2024-07"}

\item{limit}{maximum number of scenes to consider (default 20)}

\item{max_cloud}{maximum cloud cover (\%)}
}
\value{
single STAC feature (list) for the best scene
}
\description{
SEARCH SENTINEL
Find the least cloudy Sentinel-2 scene for a site and month
}
